refract(features/fracial-recognition): update useFacialRecognition hook by splitting it and follow SOLID
