+++ src/features/auth/components/FacialTrigger.tsx
@@ -2,6 +2,7 @@
 
 import { Button } from "@/components/ui/button";
 import { Facial } from "@/icons";
+import { useTheme } from "next-themes";
 import dynamic from "next/dynamic";
 
 const FacialDialogWrapper = dynamic(
@@ -12,11 +13,15 @@ const FacialDialogWrapper = dynamic(
 );
 
 const FacialTrigger = (): JSX.Element => {
+    const { theme } = useTheme();
     return (
         <FacialDialogWrapper>
             <Button variant="outline" className="h-10 w-full space-x-5">
-                <Facial className="mr-1 hidden !size-5 sm:block" /> Par
-                reconnaissance faciale
+                <Facial
+                    color={theme === "dark" ? "#ffffff" : "#0D0D0D"}
+                    className="mr-1 hidden !size-5 sm:block"
+                />{" "}
+                Par reconnaissance faciale
             </Button>
         </FacialDialogWrapper>
     );
