+++ src/features/facial/hooks/useFacialRecognition.ts
@@ -1,210 +1,54 @@
 "use client";
 
-import { useEffect, useRef, useCallback, useState } from "react";
-import * as faceapi from "face-api.js";
+import { useEffect, useRef, useState } from "react";
 import { AdminFacialRecognition } from "@prisma/client";
-import { compareFace, handleLabelFace, loadModels } from "@/lib/faceapi";
-import { createSession } from "@/lib/session";
+import { loadModels } from "@/lib/faceapi";
 import { useRouter } from "next/navigation";
+import useWebcam from "./useWebcam";
+import useFaceDetection from "./useFaceDetection";
 
-type IAuthStatus = "Authentificated" | "Unknow" | "Pending";
+export type IAuthStatus = "Authentificated" | "Unknow" | "Pending";
 
 const useFacialRecognition = (
     facialData: AdminFacialRecognition[],
     isLoadingFaces: boolean,
 ) => {
     const router = useRouter();
-    const videoRef = useRef<HTMLVideoElement>(null);
-    const canvasRef = useRef<HTMLCanvasElement>(null);
-    const streamRef = useRef<MediaStream | null>(null);
-    const isInitializedRef = useRef<boolean>(false);
-    const detectionIntervalRef = useRef<NodeJS.Timeout | null>(null);
-    const [isLoadingWebcam, setLoadingWebcam] = useState<boolean>(true);
-    const [modelsLoaded, setLoadingModel] = useState<boolean>(false);
     const isLoadingFacesRef = useRef<boolean>(isLoadingFaces);
+
+    const handleAuthResult = (status: IAuthStatus) => {
+        setAuthentificationStatus(status);
+    };
+
+    const {
+        videoRef,
+        canvasRef,
+        streamRef,
+        isInitializedRef,
+        isLoadingWebcam,
+        setLoadingWebcam,
+        closeWebcam,
+        startVideo,
+    } = useWebcam();
+
+    const { detectFace, stopDetection } = useFaceDetection(
+        facialData,
+        isLoadingFaces,
+        router,
+        videoRef,
+        canvasRef,
+        isLoadingFacesRef,
+        handleAuthResult,
+        closeWebcam,
+    );
+
+    const [modelsLoaded, setLoadingModel] = useState<boolean>(false);
     const [authentificationStatus, setAuthentificationStatus] =
         useState<IAuthStatus>("Pending");
 
     useEffect(() => {
         isLoadingFacesRef.current = isLoadingFaces;
     }, [isLoadingFaces]);
-
-    const closeWebcam = useCallback(() => {
-        if (detectionIntervalRef.current) {
-            clearInterval(detectionIntervalRef.current);
-            detectionIntervalRef.current = null;
-        }
-
-        if (streamRef.current) {
-            streamRef.current.getTracks().forEach((track) => {
-                track.stop();
-            });
-            streamRef.current = null;
-        }
-
-        if (videoRef.current) {
-            videoRef.current.pause();
-            videoRef.current.srcObject = null;
-        }
-
-        isInitializedRef.current = false;
-    }, []);
-
-    const detectFace = useCallback(() => {
-        if (detectionIntervalRef.current) {
-            clearInterval(detectionIntervalRef.current);
-        }
-
-        detectionIntervalRef.current = setInterval(async () => {
-            if (
-                canvasRef.current &&
-                videoRef.current &&
-                videoRef.current.readyState === 4
-            ) {
-                const videoWidth = videoRef.current.videoWidth;
-                const videoHeight = videoRef.current.videoHeight;
-
-                if (videoWidth === 0 || videoHeight === 0) {
-                    console.log("Video dimensions not ready");
-                    return;
-                }
-
-                const displaySize = {
-                    width: videoWidth,
-                    height: videoHeight,
-                };
-
-                canvasRef.current.width = videoWidth;
-                canvasRef.current.height = videoHeight;
-
-                faceapi.matchDimensions(canvasRef.current, displaySize);
-
-                try {
-                    const detection = await faceapi
-                        .detectSingleFace(
-                            videoRef.current,
-                            new faceapi.TinyFaceDetectorOptions(),
-                        )
-                        .withFaceLandmarks()
-                        .withFaceDescriptor();
-
-                    const resizedDetection = detection
-                        ? faceapi.resizeResults(detection, displaySize)
-                        : undefined;
-
-                    const ctx = canvasRef.current.getContext("2d");
-                    if (ctx) {
-                        ctx.clearRect(
-                            0,
-                            0,
-                            displaySize.width,
-                            displaySize.height,
-                        );
-
-                        if (resizedDetection) {
-                            faceapi.draw.drawFaceLandmarks(
-                                canvasRef.current,
-                                resizedDetection,
-                            );
-                        }
-
-                        if (resizedDetection && !isLoadingFacesRef.current) {
-                            if (!facialData || facialData.length === 0) {
-                                setTimeout(() => {
-                                    setAuthentificationStatus("Unknow");
-                                    return;
-                                }, 660);
-                            }
-
-                            try {
-                                const LabeledFaceDescriptors =
-                                    await handleLabelFace(
-                                        facialData.map((fd) => ({
-                                            image: fd.image,
-                                            username: fd.adminId,
-                                        })),
-                                    );
-
-                                const results = await compareFace(
-                                    LabeledFaceDescriptors,
-                                    resizedDetection.descriptor,
-                                );
-
-                                setTimeout(async () => {
-                                    if (results.label !== "unknown") {
-                                        setAuthentificationStatus(
-                                            "Authentificated",
-                                        );
-                                        await createSession({
-                                            method: "FACIAL_RECOGNITION",
-                                        });
-                                        router.replace("/admin/statistics");
-                                    } else {
-                                        setAuthentificationStatus("Unknow");
-                                    }
-
-                                    closeWebcam();
-                                }, 565);
-                            } catch (labelError) {
-                                console.error(
-                                    "Error processing face labels:",
-                                    labelError,
-                                );
-                            }
-                        }
-                    }
-                } catch (error) {
-                    console.error("Error during face detection:", error);
-                }
-            }
-        }, 100);
-    }, [facialData, modelsLoaded]);
-
-    const startVideo = useCallback(async () => {
-        if (isInitializedRef.current || !videoRef.current) {
-            return;
-        }
-
-        try {
-            const stream = await navigator.mediaDevices.getUserMedia({
-                video: {
-                    width: { ideal: 640 },
-                    height: { ideal: 480 },
-                },
-            });
-
-            if (!videoRef.current) {
-                stream.getTracks().forEach((track) => track.stop());
-                return;
-            }
-
-            streamRef.current = stream;
-            videoRef.current.srcObject = stream;
-
-            videoRef.current.onloadedmetadata = () => {
-                if (videoRef.current && !isInitializedRef.current) {
-                    videoRef.current
-                        .play()
-                        .then(() => {
-                            console.log("Video started playing");
-                            isInitializedRef.current = true;
-
-                            if (modelsLoaded) {
-                                setTimeout(() => {
-                                    setLoadingWebcam(false);
-                                }, 500);
-                            }
-                        })
-                        .catch((error) => {
-                            console.error("Error playing video:", error);
-                        });
-                }
-            };
-        } catch (err) {
-            console.error("Error accessing camera:", err);
-            setLoadingWebcam(false);
-        }
-    }, [modelsLoaded]);
 
     useEffect(() => {
         const loadFaceapiModels = async () => {
@@ -233,10 +77,7 @@ const useFacialRecognition = (
 
     useEffect(() => {
         if (!isLoadingWebcam && modelsLoaded && isInitializedRef.current) {
-            if (detectionIntervalRef.current) {
-                clearInterval(detectionIntervalRef.current);
-                detectionIntervalRef.current = null;
-            }
+            stopDetection();
 
             const timer = setTimeout(() => {
                 detectFace();
@@ -256,5 +97,4 @@ const useFacialRecognition = (
         detectFace,
     };
 };
-
 export default useFacialRecognition;
